refactor: split routes and api
splitted api in index.js file and routes into app.js file
